ml/fall_model/requirements.txt
google-cloud-storage>=2.16.0
pandas>=2.2.2
torch>=2.2.0
torchvision>=0.17.0
